+++ src/App.js
@@ -1,16 +1,18 @@
-import React from "react";
+import React, { useState } from "react";
 
 import Header from "./components/Header.js";
 import Sidebar from "./components/Sidebar.js";
 import Thoughts from "./components/Thoughts.js";
 
 function App() {
+	const [subreddit, setSubreddit] = useState("showerthoughts");
+
 	return (
 		<div className="App">
-			<Header />
+			<Header subreddit={subreddit} setSubreddit={setSubreddit} />
 			<div className="container">
 				<Sidebar />
-				<Thoughts />
+				<Thoughts subreddit={subreddit} />
 			</div>
 		</div>
 	);